+++ src/processor/diagnostic/report.rs
@@ -1057,6 +1057,7 @@ mod tests {
             id: "test".to_string(),
             collection_date: 0,
             runner: "test".to_string(),
+            version: None,
             uuid: "test".to_string(),
         };
         let mut report =
@@ -1081,6 +1082,7 @@ mod tests {
             id: "test".to_string(),
             collection_date: 0,
             runner: "test".to_string(),
+            version: None,
             uuid: "test".to_string(),
         };
         let mut report =
@@ -1242,6 +1244,7 @@ user: ada
             id: "test".to_string(),
             collection_date: 0,
             runner: "test".to_string(),
+            version: None,
             uuid: "test".to_string(),
         };
         let mut report =
@@ -1276,6 +1279,7 @@ user: ada
             id: "test".to_string(),
             collection_date: 0,
             runner: "test".to_string(),
+            version: None,
             uuid: "test".to_string(),
         };
         let mut report =
@@ -1475,6 +1479,7 @@ user: ada
             id: "test".to_string(),
             collection_date: 0,
             runner: "test".to_string(),
+            version: None,
             uuid: "test".to_string(),
         };
         let mut report =