+++ src/processor/kibana/node_stats/mod.rs
@@ -1,0 +1,13 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License 2.0;
+// you may not use this file except in compliance with the Elastic License 2.0.
+
+//! Kibana node stats processor
+//!
+//! Inputs:
+//! - `kibana_stats.json` (`/api/stats`) -> `metrics-kibana.node-esdiag`
+
+mod data;
+mod processor;
+
+pub use data::*;