+++ src/receiver/archive/bytes.rs
@@ -131,6 +131,15 @@ impl std::fmt::Display for ArchiveBytesReceiver {
 }
 
 impl ArchiveBytesReceiver {
+    pub async fn get_all<T>(&self) -> Result<Vec<Result<T>>>
+    where
+        T: DataSource + DeserializeOwned,
+    {
+        let mut archive = self.archive.write().await;
+        let ctx = self.source_context()?;
+        super::get_all_from_archive::<BufReader<Cursor<Bytes>>, T>(&mut archive, self.subdir.as_ref(), &ctx)
+    }
+
     pub(crate) fn clone_for_subdir(&self, work_dir: &str) -> Self {
         Self {
             archive: self.archive.clone(),